# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "STX",
        "contact": {},
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/clusters/{id}/configs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "获取集群配置列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/clusters/{id}/configs/init": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "初始化集群配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "初始化请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.InitClusterConfigsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/clusters/{id}/configs/sync-all": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "同步模板到所有节点",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "同步请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.SyncTemplateToAllNodesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/configs/normalize": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "规范化配置内容",
                "parameters": [
                    {
                        "description": "规范化请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.NormalizeConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/configs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "获取配置详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "更新配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.UpdateConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/configs/{id}/promote": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "推广配置到集群",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "推广请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.PromoteConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/configs/{id}/push": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "推送配置到节点",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "推送请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.PushConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/configs/{id}/rollback": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "回滚配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "回滚请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.RollbackConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/configs/{id}/sync": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "从集群模板同步配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "同步请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.SyncConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/configs/{id}/versions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "获取配置版本历史",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "config.InitClusterConfigsRequest": {
            "type": "object",
            "required": [
                "host_id",
                "install_dir"
            ],
            "properties": {
                "host_id": {
                    "type": "integer"
                },
                "install_dir": {
                    "type": "string"
                }
            }
        },
        "config.NormalizeConfigRequest": {
            "type": "object",
            "required": [
                "config_type",
                "content"
            ],
            "properties": {
                "config_type": {
                    "$ref": "#/definitions/github_com_LeonYoah_stx_internal_apps_config.ConfigType"
                },
                "content": {
                    "type": "string"
                }
            }
        },
        "config.PromoteConfigRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                }
            }
        },
        "config.PushConfigRequest": {
            "type": "object",
            "required": [
                "install_dir"
            ],
            "properties": {
                "install_dir": {
                    "type": "string"
                }
            }
        },
        "config.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string"
                }
            }
        },
        "config.RollbackConfigRequest": {
            "type": "object",
            "required": [
                "version"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "config.SyncConfigRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                }
            }
        },
        "config.SyncTemplateToAllNodesRequest": {
            "type": "object",
            "required": [
                "config_type"
            ],
            "properties": {
                "config_type": {
                    "type": "string"
                }
            }
        },
        "config.UpdateConfigRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                }
            }
        },
...
        "github_com_LeonYoah_stx_internal_apps_config.ConfigType": {
            "type": "string",
            "enum": [
                "seatunnel.yaml",
                "hazelcast.yaml",
                "hazelcast-client.yaml",
                "jvm_options",
                "log4j2.properties",
                "seatunnel-env.sh",
                "hazelcast-master.yaml",
                "hazelcast-worker.yaml",
                "jvm_master_options",
                "jvm_worker_options"
            ],
            "x-enum-varnames": [
                "ConfigTypeSeatunnel",
                "ConfigTypeHazelcast",
                "ConfigTypeHazelcastClient",
                "ConfigTypeJVMOptions",
                "ConfigTypeLog4j2",
                "ConfigTypeSeatunnelEnv",
                "ConfigTypeHazelcastMaster",
                "ConfigTypeHazelcastWorker",
                "ConfigTypeJVMMasterOptions",
                "ConfigTypeJVMWorkerOptions"
            ]
        },
...
    }
}